Set `density` with args in story
